# Conflicts:
#	Circus/OrderBook/InMemoryOrderBook.cs
#	Circus/OrderBook/InternalOrder.cs

+++ Circus/Util/DescendingComparer.cs
@@ -2,9 +2,9 @@ using System.Collections.Generic;
 
 namespace Circus.Util
 {
-    internal class DescendingComparer : IComparer<decimal>
+    internal class DescendingComparer : IComparer<long>
     {
-        public int Compare(decimal x, decimal y)
+        public int Compare(long x, long y)
         {
             return y.CompareTo(x);
         }
